task: cast before sub

+++ Chapter5/Exercise 5-15/AltSolution/cmp.c
@@ -27,5 +27,5 @@ int strcmpignore(const char *s1, const char *s2)
       ++s1; ++s2;
    }
 
-   return tolower(*s1) - tolower(*s2);
+   return tolower((int)*s1) - tolower((int)*s2);
 }
